src/aws-finance-mcp-server/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
python-jose==3.3.0
numpy==1.26.4
pandas>=2.2,<2.3
statsmodels>=0.14.0,<0.15
u8darts>=0.31,<0.32
river>=0.22,<0.24
pyod>=2.0.5,<2.1
